Add support for discrimated models in GooseRestRouter (#40)
If you want to have multiple types of models off the same collection,
Mongoose provides a discriminator for sub-models and inheritance. We
need to support these differently, because updating a base model with a
key only present on a discriminated model silently fails. So we allow
sending __t from the client to specify the discriminated model and an
option to support this when setting up GooseRestRouter.