Add rejection/exception handler, fix types declaration (#169)
At some point we stopped emitting .d.ts files, so types were broken
in anything consuming the package. Fix that by turning on declaration
in tsconfig.json.

Add a test to ensure using a $in query with a list of ObjectIds works.

Add an unhandled rejection handler to winston. This was causing apps to
hang when they had an unhandled rejection because winston throws a new
error (I think).