Handle population in the API
For APIs with populatePaths set to an array of strings, use those
strings to try and guess at the model being populated and add it to the
OpenAPI schema. Otherwise the schema would say it was returning a string
(for the ObjectID) but really return the whole API, messing up SDKs.

Doesn't currently support the function version of populate paths.