Update how we populate models
Ditch the option to pass a function for populatePaths. This was causing
a lot of type errors and was from the days before responseHandler was
used all over. Now versioning can be handled in responseHandler. Add a
utility function unpopulate to help with versioning (for clients
that shouldn't have a population).

Update the signature for PopulatePaths to allow limiting the fields that
are populated and an openApiComponent to allow a nice, reusable
component, which makes frontend SDKs nicer.

This is a backwards incompatible change, so apps using this will need to
change their populatePaths to match the new style. I tried to make it
work but typing and OpenAPI wasn't possible.